move fs to core
