+++ webmentions/webmentions.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/hacdias/eagle/core"
-	"github.com/hacdias/eagle/fs"
 	"github.com/hacdias/eagle/log"
 	"go.uber.org/zap"
 	"willnorris.com/go/webmention"
@@ -14,12 +13,12 @@ import (
 type Webmentions struct {
 	log      *zap.SugaredLogger
 	client   *webmention.Client
-	fs       *fs.FS
+	fs       *core.FS
 	hugo     *core.Hugo
 	notifier core.Notifier
 }
 
-func NewWebmentions(fs *fs.FS, hugo *core.Hugo, notifier core.Notifier) *Webmentions {
+func NewWebmentions(fs *core.FS, hugo *core.Hugo, notifier core.Notifier) *Webmentions {
 	return &Webmentions{
 		log: log.S().Named("webmentions"),
 		client: webmention.New(&http.Client{
